Moved Benji's stuff to new module/app. But having trouble loading template for report page
