+++ ap/attendance/urls.py
@@ -11,8 +11,6 @@ urlpatterns = [
   url(r'^rolls/team/$', views.TeamRollsView.as_view(), name='team-rolls'),
   url(r'^rolls/ypc/$', views.YPCRollsView.as_view(), name='ypc-rolls'),
   url(r'^rolls/rfid/$', views.RFIDRollsView.as_view(), name='rfid-rolls'),
-  url(r'^rolls/semi/$', views.SemiAnnualView.as_view(), name='semi-rolls'),
-  url(r'^rolls/semi/report/$', views.SemiAnnualStudyReport.as_view(), name='semi-study-report'),
   url(r'^api/rolls/finalize/$', views.finalize, name='finalize'),
   url(r'^api/rolls/rfid/(?P<trainee_id>\d+)$', views.rfid_signin),
   url(r'^api/rolls/rfid-finalize/(?P<event_id>\d+)/(?P<event_date>\d{4}-\d{2}-\d{2})$', views.rfid_finalize, name='rfid-roll-finalize'),
