+++ ap/attendance/views.py
@@ -564,53 +564,6 @@ class RFIDRollsView(TableRollsView):
     ctx['title'] = "RFID Rolls"
     return ctx
 
-class SemiAnnualView(TableRollsView):
-  def get_context_data(self, **kwargs):
-    kwargs['selected_date'] = CURRENT_TERM.get_date(19, 3)
-    kwargs['trainees'] = Trainee.objects.filter(pk=self.request.user.id)
-    kwargs['event_type'] = 'S'
-    kwargs['monitor'] = 'HC'
-    ctx = super(SemiAnnualView, self).get_context_data(**kwargs)
-    ctx['title'] = "Semi-annual Study Attendance"
-    return ctx
-
-class SemiAnnualStudyReport(ListView):
-  model = Roll
-  template_name = 'attendance/semi-annual_study_attendance.html'
-
-  def get_context_data(self, **kwargs):
-    ctx = super(SemiAnnualStudyReport, self).get_context_data(**kwargs)
-    ctx['title'] = 'Semi-annual Study Attendance Report'
-
-    evs = Schedule.objects.get(weeks=19).events.filter(type='S')
-    rolls = Roll.objects.filter(event__in=evs).order_by('trainee__lastname', 'trainee__firstname')
-
-    ctx['headers'] = ['Name', 'Term', 'Location', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
-    roll_dict = []
-    for t in Trainee.objects.filter(is_active=True):
-      t_rolls = rolls.filter(trainee=t)
-      t_dict = {
-       'Name': t.full_name2,
-       'Term': t.current_term,
-       'Tue': ' ',
-       'Wed': ' ',
-       'Thu': ' ',
-       'Fri': ' ',
-       'Sat': ' ',
-      }
-
-      # TODO add location
-      t_dict['Location'] = 'TC'
-
-      for r in t_rolls:
-        w_day = r.date.strftime("%a")
-        t_dict[w_day] = r.status
-
-      roll_dict.append(t_dict)
-
-    ctx['rolls'] = roll_dict
-    return ctx
-
 
 class RollViewSet(BulkModelViewSet):
   queryset = Roll.objects.all()
